+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added Clear button to event filter bar [#4041](https://github.com/rokwire/illinois-app/issues/4041).
 
 ## [5.2.8] - 2024-05-10
 ### Changed